requirements.txt
mediapipe>=0.10.0
opencv-python>=4.8.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
pyyaml>=6.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
